Inspect for series
